Patch from Thomas Leonard:
Fixed race condition in the boot-comm system

When queuing an event in another vat on the same machine, we tried to construct
a BootRefHandler for the promise as the promise was being resolved in another
thread. This could lead to various race conditions, including errors such as
"OldFarRef may only be smashed", "ViciousCycleException" and deadlock).

Now, we create the BootRefHandler for the promise of the result before adding
it to the queue.

This might also mean that a number of "this method must be thread-safe"
requirements can be dropped.

git-svn-id: svn://svn.synchrona.org/erights/e/trunk@718 bdfa04cf-0718-0410-970a-89734667f18f