Serialize Switchable refs correctly
Java serialization gives each object the chance to replace itself with another
object. ConstMapImpl uses this to change itself into a ConstMap.fromColumns() call.
However, when serializing a resolved SwitchableRef Java only gives the SwitchableRef
this option, not the underlying object, so we should call it manually.

Also, if we do get a raw ConstMapImpl object, throw an expcetion immediately
rather than constructing an invalid object (with myTable == null).

Patch from the University of Southampton IT Innovation Centre.

git-svn-id: svn://svn.synchrona.org/erights/e/trunk@740 bdfa04cf-0718-0410-970a-89734667f18f