Use SecureRandom instead of ESecureRandom everywhere
This allows other implementations to be used, and is especially useful for
testing (where a faster but less secure version is desirable).

ESecureRandom.nextSwiss was the only extra method being used. All callers have
been changed to use the (pre-existing) nextSwiss method on SwissTable instead.

Patch from the University of Southampton IT Innovation Centre.

git-svn-id: svn://svn.synchrona.org/erights/e/trunk@754 bdfa04cf-0718-0410-970a-89734667f18f