Convenience methods for traceln
You can now pass any object to traceln.run/1, not just a string, and it
will be converted to a string. If the object is a Throwable, it will
also give the E stack-trace.

There's also a new run/2 that takes a string and a throwable.

Patch from the University of Southampton IT Innovation Centre.

git-svn-id: svn://svn.synchrona.org/erights/e/trunk@799 bdfa04cf-0718-0410-970a-89734667f18f