deSubgraphKit: use separate tables for temps and unscope.
Previously, deSubgraphKit conflated the table of subgraph exits (unscope) with the table mapping objects to temp indexes. This permits the passed-in unscope to refer to some arbitrary temp in the graph being serialized by containing an integer rather than a string. This is a POLA violation and fails to catch mistakes.

Discussion: http://www.eros-os.org/pipermail/e-lang/2009-May/013154.html

git-svn-id: svn://svn.synchrona.org/erights/e/trunk@802 bdfa04cf-0718-0410-970a-89734667f18f